@@ -1,19 +1,19 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 20
-    buildToolsVersion "20.0.0"
+    compileSdkVersion 29
+    buildToolsVersion "29.0.3"
 
     defaultConfig {
         applicationId "com.danyalvarez.primicia"
-        minSdkVersion 8
-        targetSdkVersion 20
+        minSdkVersion 19
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
     }
     buildTypes {
         release {
-            runProguard false
+            minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
@@ -24,16 +24,16 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile files('libs/twitter4j-core-4.0.2.jar')
-    compile files('libs/twitter4j-async-4.0.2.jar')
-    compile files('libs/picasso-2.3.4.jar')
-    //    compile 'com.github.manuelpeinado.fadingactionbar:fadingactionbar-abc:3.1.2'
-    compile('com.github.manuelpeinado.fadingactionbar:fadingactionbar-abc:3.1.2') {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation files('libs/twitter4j-core-4.0.2.jar')
+//    implementation files('libs/twitter4j-async-4.0.2.jar')
+    implementation files('libs/picasso-2.3.4.jar')
+    //    implementation 'com.github.manuelpeinado.fadingactionbar:fadingactionbar-abc:3.1.2'
+    implementation('com.github.manuelpeinado.fadingactionbar:fadingactionbar-abc:3.1.2') {
         exclude group: 'com.android.support', module: 'appcompat-v7'
     }
-    compile files('libs/Parse-1.5.1.jar')
-    compile 'com.android.support:appcompat-v7:20.0.0'
-    compile 'com.android.support:support-v4:20.0.0'
-    compile 'com.makeramen:roundedimageview:1.3.0'
+//    implementation files('libs/Parse-1.5.1.jar')
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.makeramen:roundedimageview:1.3.0'
 }
